WIP on inhouse motifbr
I committed at this stage to mark a tipical problem that arises with my (bad) way
of working and snakemake: I have an input file for the annotate_motif_divided
rule that is done in another dir (without a subworkflow, which I tend to avoid
when I'm running running running and I do not want to retrigger making something by mistake)
but which name matches a rule also in this Snakefile, which obviously then fails due
to missing input. I could solve it using a subworkflow or changing that input file
to a params...I do not love either solutions.