WIP on knitting overlap and inhouse motifbr
My comment on 5137ca5658572d124915e857608ed75e40ee5e63 was wrong, I simply
had a wrong path to the matrix file: it tries to build it only if it does not
find the file, apparently. Commented out the subworkflow thingy: having it in conf.sk
satisfy me, even if the risk of re-doing something is there.